pyproject.toml
[project]
name = "novel-to-toon"
version = "0.1.0"
description = "Agentic workflow for converting novels to webtoon panels"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.75.0",
    "fal-client>=0.5.0",
    "google-genai>=1.0.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
    "claude-agent-sdk>=0.1.33",
]
